Added custom props for divider and dateInput (#221)
