+++ packages/vue3/src/components/DateInput/DatePicker.vue
@@ -4,11 +4,9 @@ import { DateUtils } from '@youcan/ui-core';
 import { computed, ref } from 'vue';
 import Day from './Internal/Day.vue';
 import MonthSwitcher from './Internal/MonthSwitcher.vue';
-import type { DateInputValue } from './types';
+import type { DateInputValue, DatePickerProps } from './types';
 
-const props = defineProps<{
-  modelValue: DateInputValue
-}>();
+const props = defineProps<DatePickerProps>();
 
 const emit = defineEmits(['update:modelValue']);
