+++ packages/vue3/src/components/DateInput/types.ts
@@ -1,8 +1,12 @@
-export type DayEdge = 'start' | 'end' | 'middle' | 'both' | 'none';
-
-export type DayDisplayFormat = 'numeric' | '2-letters';
-
 export interface DateInputValue {
   start?: Date
   end?: Date
+}
+export interface DateInputProps {
+  modelValue: DateInputValue
+  disabled?: boolean
+}
+
+export interface DatePickerProps {
+  modelValue: DateInputValue
 }
