+++ packages/vue3/src/components/Divider/types.ts
@@ -1,3 +1,11 @@
 export type Orientation = 'horizontal' | 'vertical';
+
 export type Style = 'solid' | 'dotted' | 'dashed';
+
 export type Thickness = 'light' | 'regular' | 'thick';
+
+export interface DividerProps {
+  orientation?: Orientation
+  borderStyle?: Style
+  thickness?: Thickness
+}
